oop session 6 :

+++ src/src/Summation.php
@@ -4,7 +4,7 @@ namespace App;
 
 class Summation
 {
-    public function sum(int $a, int $b): int
+    public function sum(int $a, int $b): float
     {
         return $a + $b;
     }
